feat: Enhance calendar functionality with user-specific data retrieval and new calendar page structure
